Clean up old stories directory
