+++ package.json
@@ -72,7 +72,6 @@
     "@types/node": "^16.0.0",
     "@types/react": "^18.2.12",
     "@types/react-dom": "^18.2.5",
-    "@types/react-syntax-highlighter": "^15.5.5",
     "autoprefixer": "^9.4.7",
     "babel-jest": "^29.6.1",
     "camelcase": "^7.0.1",
@@ -103,7 +102,6 @@
     "react-helmet-async": "^1.3.0",
     "react-i18next": "^13.3.0",
     "react-router-dom": "^6.3.0",
-    "react-syntax-highlighter": "^15.5.0",
     "react-textarea-autosize": "^8.5.2",
     "regenerator-runtime": "^0.13.1",
     "sass": "^1.62.0",
