--- stories/index.ts
@@ -1,12 +1,0 @@
-import '@fontsource/source-sans-pro/index.css';
-import '@fontsource/source-sans-pro/400-italic.css';
-import '@fontsource/source-sans-pro/300.css';
-import '@fontsource/source-sans-pro/300-italic.css';
-import '@fontsource/source-sans-pro/600.css';
-import '@fontsource/source-sans-pro/700.css';
-import '@fontsource/source-code-pro/index.css';
-import '@fontsource/source-code-pro/400-italic.css';
-import '@fontsource/source-code-pro/700.css';
-import '@fontsource/source-serif-pro/index.css';
-import '@fontsource/source-serif-pro/400-italic.css';
-import '@fontsource/source-serif-pro/700.css';
