+++ stories/wrappers/index.ts
@@ -1,3 +1,2 @@
 export { default as StoryBody } from './StoryBody';
 export { default as StoryIntro } from './StoryIntro';
-export { default as IconList } from './IconList';
